+++ lib/features/theory/domain/services/scale_degree_function.dart
@@ -86,6 +86,10 @@ _Family _family(ScaleKind kind) {
     ScaleKind.majorPentatonic ||
     ScaleKind.minorPentatonic ||
     ScaleKind.majorBlues ||
-    ScaleKind.minorBlues => _Family.modal,
+    ScaleKind.minorBlues ||
+    ScaleKind.wholeTone ||
+    ScaleKind.augmented ||
+    ScaleKind.diminishedWholeHalf ||
+    ScaleKind.diminishedHalfWhole => _Family.modal,
   };
 }